Add boolean constraints built from predicates
A constraint is essentially a predicate, just one that the solver must
satisfy. Therefore, constraints and predicates have been combined,
allowing further boolean predicates to be built from any type of
predicate. This adds to the expressiveness of the system primarily
by enabling predicates to be ORed - separate constraints are essentially
ANDed already.