Model events as compound variables
Events have a start, end and duration. These have a relationship to each
other that can be defined using constraints so that when on or more of
the variables is fixed, the others can be inferred.

More complex relationships between events can be defined in terms of
constraints between their constituent variables, assuming that all
constraints hold.