Split solver into separate module
The solver can only run on linux due to its
dependency on the minizinc solver. However, the
constraints themselves can be built on any
platform. Separating the problem definition from
the solver means that a browser-based application
could build the problem, then post to a server to
actually solve it.